+++ x/e2ee/types/genesis.go
@@ -8,5 +8,10 @@ func DefaultGenesis() *GenesisState {
 // Validate performs basic genesis state validation returning an error upon any
 // failure.
 func (gs GenesisState) Validate() error {
+	for _, key := range gs.Keys {
+		if err := key.Validate(); err != nil {
+			return err
+		}
+	}
 	return nil
 }