+++ CHANGELOG.md
@@ -19,6 +19,10 @@
 
 * (rpc) [#1397](https://github.com/crypto-org-chain/cronos/pull/1397) Avoid panic on invalid elasticity_multiplier.
 
+### Features
+
+*  [#1406](https://github.com/crypto-org-chain/cronos/pull/1406) Add set-encryption-key for encryption module.
+
 *April 8, 2024*
 
 ## v1.2.0-rc1